Slight refactoring and improvements
