+++ ydb/core/viewer/viewer_ut.cpp
@@ -1,9 +1,9 @@
+#include "ut/ut_utils.h"
 #include <library/cpp/testing/unittest/registar.h>
 #include <library/cpp/testing/unittest/tests_data.h>
 #include <ydb/library/actors/interconnect/interconnect.h>
 #include <ydb/library/actors/helpers/selfping_actor.h>
 #include <library/cpp/http/misc/httpcodes.h>
-#include <library/cpp/http/simple/http_client.h>
 #include <library/cpp/json/json_value.h>
 #include <library/cpp/json/json_reader.h>
 #include <util/stream/null.h>
@@ -20,9 +20,9 @@
 #include <ydb/core/kqp/common/kqp.h>
 #include <ydb/core/testlib/test_client.h>
 #include <ydb/core/testlib/tenant_runtime.h>
-#include <ydb/core/testlib/test_pq_client.h>
 
 #include <ydb/public/lib/deprecated/kicli/kicli.h>
+
 #include <ydb/core/node_whiteboard/node_whiteboard.h>
 #include <ydb/library/actors/core/interconnect.h>
 #include <util/string/builder.h>
@@ -35,6 +35,7 @@ using namespace NKikimrWhiteboard;
 using namespace NSchemeShard;
 using namespace Tests;
 using namespace NMonitoring;
+using namespace NKikimr::NViewerTests;
 using TNavigate = NSchemeCache::TSchemeCacheNavigate;
 
 #ifdef NDEBUG
@@ -193,51 +194,6 @@ Y_UNIT_TEST_SUITE(Viewer) {
         UNIT_ASSERT_VALUES_EQUAL(result.PDiskStateInfoSize(), 100000);
         Ctest << "Data has merged" << Endl;
     }
-
-    struct THttpRequest : NMonitoring::IHttpRequest {
-        HTTP_METHOD Method;
-        TCgiParameters CgiParameters;
-        THttpHeaders HttpHeaders;
-        TString PostContent;
-
-        THttpRequest(HTTP_METHOD method)
-            : Method(method)
-        {}
-
-        ~THttpRequest() {}
-
-        const char* GetURI() const override {
-            return "";
-        }
-
-        const char* GetPath() const override {
-            return "";
-        }
-
-        const TCgiParameters& GetParams() const override {
-            return CgiParameters;
-        }
-
-        const TCgiParameters& GetPostParams() const override {
-            return CgiParameters;
-        }
-
-        TStringBuf GetPostContent() const override {
-            return PostContent;
-        }
-
-        HTTP_METHOD GetMethod() const override {
-            return Method;
-        }
-
-        const THttpHeaders& GetHeaders() const override {
-            return HttpHeaders;
-        }
-
-        TString GetRemoteAddr() const override {
-            return TString();
-        }
-    };
 
     class TMonPage: public IMonPage {
     public:
@@ -585,18 +541,6 @@ Y_UNIT_TEST_SUITE(Viewer) {
         const TKeepAliveHttpClient::THttpCode statusCode = httpClient.DoPost("/viewer/query?timeout=600000", NJson::WriteJson(jsonRequest, false), &responseStream, headers);
         UNIT_ASSERT_EQUAL(statusCode, HTTP_OK);
         return NJson::ReadJsonTree(&responseStream, /* throwOnError = */ true);
-    }
-
-    void WaitForHttpReady(TKeepAliveHttpClient& client) {
-        for (int retries = 0;; ++retries) {
-            UNIT_ASSERT(retries < 100);
-            TStringStream responseStream;
-            const TKeepAliveHttpClient::THttpCode statusCode = client.DoGet("/viewer/simple_counter?max_counter=1&period=100", &responseStream);
-            const TString response = responseStream.ReadAll();
-            if (statusCode == HTTP_OK) {
-                break;
-            }
-        }
     }
 
     void GrantConnect(TClient& client) {
@@ -2117,77 +2061,4 @@ Y_UNIT_TEST_SUITE(Viewer) {
         UNIT_ASSERT_EQUAL_C(statusCode, HTTP_BAD_REQUEST, statusCode << ": " << response);
         UNIT_ASSERT_C(response.StartsWith("Conversion error"), response);
     }
-
-
-    Y_UNIT_TEST(GetTopicDataTest) {
-        TPortManager tp;
-        ui16 port = tp.GetPort(2134);
-        ui16 grpcPort = tp.GetPort(2135);
-        ui16 monPort = tp.GetPort(8765);
-
-        auto settings = NKikimr::NPersQueueTests::PQSettings(port, 1);
-        settings.PQConfig.MutableQuotingConfig()->SetEnableQuoting(false);
-        settings.PQConfig.SetTopicsAreFirstClassCitizen(true);
-
-        settings.InitKikimrRunConfig()
-                .SetNodeCount(1)
-                .SetUseRealThreads(true)
-                .SetDomainName("Root")
-                .SetMonitoringPortOffset(monPort, true);
-
-        auto grpcSettings = NYdbGrpc::TServerOptions().SetHost("[::1]").SetPort(grpcPort);
-        TServer server{settings};
-        server.EnableGRpc(grpcSettings);
-
-        auto client = MakeHolder<NKikimr::NPersQueueTests::TFlatMsgBusPQClient>(settings, grpcPort);
-        client->InitRoot();
-        client->InitSourceIds();
-        client->CheckClustersList(server.GetRuntime());
-        NYdb::TDriverConfig driverCfg;
-        TString topicPath = "/Root/topic1";
-        driverCfg.SetEndpoint(TStringBuilder() << "localhost:" << grpcPort)
-                 .SetLog(std::unique_ptr<TLogBackend>(CreateLogBackend("cerr", ELogPriority::TLOG_DEBUG).Release()));
-
-        std::shared_ptr<NYdb::TDriver> ydbDriver(new NYdb::TDriver(driverCfg));
-        auto topicClient = NYdb::NTopic::TTopicClient(*ydbDriver);
-
-        auto res = topicClient.CreateTopic(topicPath).GetValueSync();
-        UNIT_ASSERT_C(res.IsSuccess(), res.GetIssues().ToString());
-
-        NYdb::NTopic::TWriteSessionSettings wsSettings;
-        wsSettings.Path(topicPath);
-        wsSettings.ProducerId("12345");
-        Cerr << "Write data\n";
-        auto writer = topicClient.CreateSimpleBlockingWriteSession(wsSettings);
-        for (auto i = 0u; i < 20; ++i) {
-            writer->Write(TStringBuilder() << "Message " << i);
-        }
-        Cerr << "Close writer\n";
-        writer->Close();
-        Cerr << "Write data - done\n";
-
-
-        TKeepAliveHttpClient httpClient("localhost", monPort);
-        WaitForHttpReady(httpClient);
-        TStringStream responseStream;
-        NHttp::TUrlParameters params{"/viewer/get_topic_data"};
-        params["topic_path"] = topicPath;
-        params["database"] = "Root";
-
-
-        TKeepAliveHttpClient::THeaders headers;
-        headers["Accept"] = "application/json";
-
-        THttpRequest httpReq(HTTP_METHOD_GET);
-        httpReq.HttpHeaders.AddHeader("Accept", "application/json");
-
-        const TKeepAliveHttpClient::THttpCode statusCode = httpClient.DoGet(params.Render(), &responseStream, headers);
-        const TString response = responseStream.ReadAll();
-        UNIT_ASSERT_EQUAL_C(statusCode, HTTP_OK, statusCode << ": " << response);
-        NJson::TJsonReaderConfig jsonCfg;
-        NJson::TJsonValue json;
-        NJson::ReadJsonTree(response, &jsonCfg, &json, /* throwOnError = */ true);
-    }
-
-
 }
