+++ ydb/core/viewer/ya.make
@@ -72,7 +72,7 @@ SRCS(
     viewer_describe.h
     viewer_describe_topic.h
     viewer_feature_flags.h
-    viewer_get_topic_data.h
+    viewer_get_topic_data.cpp
     viewer_graph.h
     viewer_healthcheck.h
     viewer_helper.h
